pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "setuptools-scm>=6.2"]
build-backend = "setuptools.build_meta"

# --- Project Metadata (PEP 621) ---
[project]
name = "natural-pdf"
# Version is now managed dynamically by setuptools-scm based on Git tags
dynamic = ["version"]
authors = [
    { name = "Jonathan Soma", email = "jonathan.soma@gmail.com" },
]
description = "A more intuitive interface for working with PDFs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "scikit-learn",
    "markdown",
    "pandas",
    "pdfplumber>=0.11.7",
    "colormath2",
    "pillow",
    "colour",
    "numpy",
    "urllib3",
    "tqdm",
    "pydantic",
    "jenkspy",
    "scipy",
    "ipywidgets>=7.0.0",
    "python-bidi",
    "matplotlib",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-xdist",
    "setuptools",
]

dev = [
    "black",
    "isort",
    "mypy",
    "pytest",
    "pytest-xdist",
    "nox",
    "nox-uv",
    "build",
    "uv",
    "pipdeptree",
    "nbformat",
    "jupytext",
    "nbclient==0.10.2",
    "jupyter_core==5.7.2",
    "ipykernel",
    "pre-commit",
    "setuptools",
]

all = [
    "natural-pdf[ocr-export]",
    "natural-pdf[deskew]",
    "natural-pdf[test]",
    "natural-pdf[search]",
    "natural-pdf[export-extras]",
    "natural-pdf[ai]"
]

deskew = [
    "deskew>=1.5",
    "img2pdf"
]

search = [
    "lancedb",
    "pyarrow",
]

ocr-export = [
    "pikepdf",
]

export-extras = [
    "jupytext",
    "nbformat",
]

ai = [
    "sentence-transformers",
    "torch",
    "torchvision",
    "transformers[sentencepiece]",
    "huggingface_hub>=0.29.3",
    "timm",
    "doclayout_yolo",
    "easyocr",
    "openai",
]

[project.urls]
Homepage = "https://github.com/jsoma/natural-pdf"
Repository = "https://github.com/jsoma/natural-pdf"

# --- Tool Configurations ---

[tool.setuptools_scm]
local_scheme = "no-local-version"

[tool.black]
line-length = 100
target-version = ['py37']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
filterwarnings = [
    # Ignore specific deprecation warnings from dependencies
    "ignore:The distutils package is deprecated:DeprecationWarning",
    "ignore:Support for class-based `config` is deprecated",
    "ignore:deprecated:DeprecationWarning:docling_core.types.doc.document",
]
markers = [
    "tutorial: tests that execute markdown tutorials to populate notebooks",
    "slow: tests that take significantly longer or exercise heavyweight resources",
    "ocr: tests that require OCR or vision models",
    "network: tests that rely on external network resources",
    "qa: tests that cover document QA functionality",
    "qa_remote: QA tests that rely on hosted LLM endpoints",
    "qa_local: QA tests that require large local QA models or embeddings",
    "optional_deps: tests that require optional or extra dependencies",
]

[tool.setuptools]
include-package-data = false

[tool.setuptools.package-data]
"natural_pdf.exporters" = ["data/*.ttf", "data/*"]
"natural_pdf" = ["templates/*.html"]

[tool.setuptools.packages.find]
where = ["."]
exclude = [
    "tests", "tests.*",
    "build", "dist", "*.egg-info",
    "docs", "docs.*",
    ".nox", ".venv", "env", "venv"
]

[project.scripts]
# natural-pdf command exposes helper CLI
natural-pdf = "natural_pdf.cli:main"
npdf = "natural_pdf.cli:main"
